requirements.txt
Phidget22==1.0.0
numpy
pygame
git+https://github.com/dpallot/simple-websocket-server.git
pyserial